requirements.txt
appdirs==1.4.4
asgiref==3.3.1
click==7.1.2
dataclasses==0.8
Django==3.1.7
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
importlib-metadata==3.7.3
Markdown==3.3.4
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.8.1
pycodestyle==2.7.0
pyflakes==2.3.0
PyJWT==1.7.1
pytz==2021.1
regex==2021.3.17
sqlparse==0.4.1
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
zipp==3.4.1
